Add save-to-file(s) feature, using ',vf'
This is very similar to append-to-file ',vp'.
However when saving (using ',vf') you don't need to pass a filename.
Instead, each message is saved to a different file.
Message subject is used as filename. Invalid characters get
replaced by underscores.

This is most useful for me to save patch messages.

TODO: Ask for confirmation? Ask for directory?
Use message author also as filename?

Signed-off-by: Ezequiel Garcia <elezegarcia@gmail.com>